\docType{package}
\name{eirm-package}
\alias{eirm-package}
\title{Profile Analysis of Multivariate Data in R}
\description{
  The package eirm provides a set of tools to analyze and visualize
  dichotomous and polytomous item responses using the explanatory item
  response modeling (EIRM). The generalized linear mixed modeling (GLMM)
  framework is used for examining the impact of item-related and person-related
  variables on responses. The items are repeated measures that  are nested underpersons.
  The model details can be found at de Boeck and Wilson's (2004) Explanatory Item Response Models -
  A Generalized Linear and Nonlinear Approach <DOI: 10.1007/978-1-4757-3990-9>.
}
\author{
  Okan Bulut \email{bulut@ualberta.ca}
}
\references{
  De Boeck, P., & Wilson, M. (2004). Explanatory item response models: A generalized linear
  and nonlinear approach. New York: Springer-Verlag.
}
\keyword{package}
