% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eirmShiny.R
\name{eirmShiny}
\alias{eirmShiny}
\title{Graphical user interface (GUI) of the eirm function}
\usage{
eirmShiny()
}
\description{
An interactive Shiny application for running the eirm function. The application allows users
to import the (long-format) response data, define the response variable and predictors, and
run the estimation, and produce the output, as well as the item-person plot, on the screen.
}
\examples{
\dontrun{
eirmShiny()
}
}
