% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simEIRM.R
\name{simEIRM}
\alias{simEIRM}
\title{Simulate data for explanatory item response modeling}
\usage{
simEIRM(
  nitem,
  nperson,
  theta.mean = 0,
  theta.sd = 1,
  difficulty.mean = 0,
  difficulty.sd = 0.8,
  group.size = NULL,
  itemtype.size = NULL,
  group.effect = c(0, 0),
  itemtype.effect = c(0, 0),
  dif = 0,
  seed = NULL
)
}
\arguments{
\item{nitem}{Number of items to be simulated.}

\item{nperson}{Number of persons to be simulated.}

\item{theta.mean}{Mean of a normal distribution to be used for generating theta values (default is 0).}

\item{theta.sd}{Standard deviation of a normal distribution to be used for generating theta values (default is 1).}

\item{difficulty.mean}{Mean of a normal distribution to be used for generating item difficulty values (default is 0).}

\item{difficulty.sd}{Standard deviation of a normal distribution to be used for generating item difficulty values (default is 0.8).}

\item{group.size}{The number of persons for each group (i.e., Group 0 vs. Group 1). The sum of these numbers
should be equal to nperson. If NULL, the first half of the persons becomes Group 0 and the other half becomes Group 1.}

\item{itemtype.size}{The number of items for each itemtype (i.e., itemtype 0 vs. itemtype 1). The sum of these numbers
should be equal to nitem. If NULL, the first half of the items becomes itemtype 0 and the other half becomes itemtype 1.}

\item{group.effect}{The magnitudes of person group effect  (in logits) for Group 0 and Group 1, respectively.}

\item{itemtype.effect}{The magnitudes of item type effect (in logits) for itemtype 0 and item type 1, respectively.}

\item{dif}{The magnitude of differential item functioning (DIF) for the items. Either a single value to create DIF
for all items or a vector with the length of nitem that specifies the amount of DIF for each item.}

\item{seed}{A number to be used as a seed in the random number generator. Especially in simulation studies, setting the
seed would ensure reproducibility of the results.}
}
\value{
A long-format response data set is returned, with the following components:
\itemize{
\item person: A pseudo person ID from 1 to nperson.
\item item: A pseudo item ID from 1 to nitem.
\item theta: Ability (i.e., theta) values used for generating responses.
\item difficulty: Item difficulty (i.e., b parameter) values used for generating responses.
\item group: Person groups (either 0 or 1).
\item group.effect: The magnitude of person group effect used for generating responses.
\item itemtype: Item groups (either 0 or 1).
\item itemtype.effect: The magnitude of itemtype effect used for generating responses.
\item dif: The magnitude of DIF used for generating responses.
\item response: Dichotomous responses generated based on theta and difficulty.
}
}
\description{
This function simulates a dichotomous response data set based on item-level and person-level predictors.
There are two predictors: an item-level (itemtype: 0 or 1) and a person-level (group: 0 or 1).
Dichotomous responses are generated for user-defined numbers of items and persons. This simple data set can
be used for simulating and testing different EIRM model with item, person, and item-by-person predictors.
}
\examples{
# A Linear Logistic Test Model (LLTM) with item group effects with 10 items and 1000 persons
# First 4 items will be itemtype = 0 and the rest will be itemtype = 1
# Items with itemtype = 1 will be 0.3 logit more difficult.
data.lltm <- simEIRM(nitem = 10, nperson = 1000, itemtype.size = c(4, 6),
itemtype.effect = c(0, 0.3))

# A latent regression model (LRM) with person group effects with 10 items and 1000 persons
# Half of the persons will be group = 0 and the other half will be group = 1 (Default)
# The average ability for persons with group = 1 will be 0.5 logit higher.
data.lrm <- simEIRM(nitem = 10, nperson = 1000, group.effect = c(0, 0.5))

# A differential item functioning (DIF) scenario with 15 items and 1000 persons
# The last five items will have DIF with 0.3 logit. Group 0 will have 700 persons (reference group)
# and Group 1 will have 300 persons (focal group)
data.dif <- simEIRM(nitem = 15, nperson = 1000, group.size = c(700, 300),
dif = c(rep(0, 10), rep(0.3, 5)))
}
